requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
apiclient==1.0.4
async-timeout==4.0.2
attrs==25.1.0
auto-py-to-exe==2.45.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
bs4==0.0.2
cachetools==5.5.1
certifi==2023.7.22
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.8
cloudscraper==1.2.71
colorama==0.4.6
cryptography==44.0.1
discord.py==2.2.2
Eel==0.18.1
fake-headers==1.0.2
fake-useragent==2.0.3
fastapi==0.115.11
Flask==3.1.0
flask-cors==5.0.1
fpdf==1.7.2
frozenlist==1.3.3
future==1.0.0
geographiclib==2.0
geopy==2.4.1
gevent==24.11.1
gevent-websocket==0.10.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.0.4
numpy==2.2.3
oauth2client==4.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
pefile==2023.2.7
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
PyMuPDF==1.25.3
pyparsing==3.2.1
pypdf==5.3.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PySimpleGUI~=5.0.8
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytrends==4.9.2
pytz==2025.1
pywin32-ctypes==0.2.3
reportlab==4.3.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
schedule==1.2.2
scipy==1.15.1
selenium==4.28.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.39
starlette==0.46.1
stripe==12.0.1
traits==7.0.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
ua-generator==2.0.0
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.34.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.8.2
zope.event==5.0
zope.interface==7.2
python-dotenv==1.0.0
